refactor(#17): move utils to a file
